+++ src/io/segment/android/Analytics.java
@@ -41,7 +41,7 @@ import android.text.TextUtils;
 
 public class Analytics {
 	
-	public static final String VERSION = "0.4.4";
+	public static final String VERSION = "0.5.0";
 	
 	private static AnalyticsStatistics statistics;